term_test.go
package terminal

import (
	"io"
	"strings"
	"testing"
)

func extractStr(t *Term, x0, x1, row int) string {
	var s []rune
	for i := x0; i <= x1; i++ {
		s = append(s, t.lines[row][i].c)
	}
	return string(s)
}

func TestPlainChars(t *testing.T) {
	term := New(80, 24, nil)
	expected := "Hello world!"
	_, err := term.Write([]byte(expected))
	if err != nil && err != io.EOF {
		t.Fatal(err)
	}
	actual := extractStr(term, 0, len(expected)-1, 0)
	if expected != actual {
		t.Fatal(actual)
	}
}

func TestNewline(t *testing.T) {
	term := New(80, 24, nil)
	expected := "Hello world!\n...and more."
	_, err := term.Write([]byte("\033[20h")) // set CRLF mode
	if err != nil && err != io.EOF {
		t.Fatal(err)
	}
	_, err = term.Write([]byte(expected))
	if err != nil && err != io.EOF {
		t.Fatal(err)
	}

	split := strings.Split(expected, "\n")
	actual := extractStr(term, 0, len(split[0])-1, 0)
	actual += "\n"
	actual += extractStr(term, 0, len(split[1])-1, 1)
	if expected != actual {
		t.Fatal(actual)
	}
}
